+++ uikit/src/main/java/com/sendbird/uikit/consts/StringSet.kt
@@ -132,6 +132,7 @@ object StringSet {
     const val EVENT_MESSAGE_UPDATED = "EVENT_MESSAGE_UPDATED"
     const val EVENT_TYPING_STATUS_UPDATED = "EVENT_TYPING_STATUS_UPDATED"
     const val EVENT_MESSAGE_TEMPLATE_UPDATED = "EVENT_MESSAGE_TEMPLATE_UPDATED"
+    const val EVENT_USER_MARKED_UNREAD = "EVENT_USER_MARKED_UNREAD"
     const val INVALID_URL = "INVALID_URL"
     const val photo = "photo"
     const val photos = "photos"